It's not a requirement to run the script as root
The script now can run without being root, but if there's the need to
install additional packages, it will stop and give the user the command
needed to install the missing packages. If run as root, it will try to
install the missing packages automatically.

Once the missing packages (if any) are installed, there is no need to
run as root anymore.